requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
distlib==0.3.7
Django==3.2.20
django-admin-honeypot==1.1.0
django-cors-headers==4.2.0
django-filter==23.2
django-redis==5.3.0
djangorestframework==3.14.0
Faker==19.2.0
filelock==3.12.2
fordev==1.0.5
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.3
Pillow==10.0.0
platformdirs==3.9.1
Pygments==2.15.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
validate-docbr==1.10.0
virtualenv==20.24.0
zipp==3.16.2
